Refactor code to follow SLAP and pass the checkstyle
