+++ src/main/java/duke/parser/UserInputParser.java
@@ -1,0 +1,2 @@
+package duke.parser;public class UserInputParser {
+}
